Make a "Docker" build and replace data to env vars
Adapt a code to a "Docker Compose". Remove a `remotePatterns` key from a
config. Update packages. Change a `README.md` file. Fix a setting of a
`chatId` cookie. Adapt a code to a change of type of a chat ID from an
integer to a string. Move a code from a `ChatCommunicator` class to
`widget.tsx` and simplify. Add an `ulid` library instead of an own
simple implementation. Update `useEffect` dependencies in a
`textbox.tsx` file.